pyproject.toml
[tool.poetry]
name = "django-garnett"
version = "0.4.8"
description = "Simple translatable Django fields"
authors = ["Aristotle Metadata Enterprises"]
license = "BSD-3-Clause"
readme = "README.md"
packages = [
    {include = "garnett"}
]

[tool.poetry.dependencies]
python = "^3.7"
django = "~3.1.0"
langcodes = "~3.3.0"
language-data = "~1.0.1"

[tool.poetry.dev-dependencies]
black = "~20.8b1"
coverage = "~5.3.1"
dj-database-url = "~0.5.0"
psycopg2 = "~2.8.6"
mock = "~4.0.3"
djangorestframework = "~3.12.4"
drf_yasg = "~1.20.0"
django-reversion = "~4.0.1"
django-reversion-compare = "~0.14.1"
django-filter = ">=21.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '''
^/tests/library_app/migrations/*  # exclude test app migrations
'''
